+++ forui_hooks/lib/forui_hooks.dart
@@ -1,4 +1,6 @@
 export '/src/accordion_controller_hook.dart';
 export '/src/calendar_controller_hook.dart';
+export '/src/popover_controller_hook.dart';
+export '/src/resizable_controller_hook.dart';
 export '/src/tab_controller_hook.dart';
 export '/src/tooltip_controller_hook.dart';